Implement nesting for Methods with same name.
This would all be lovely if I hadn't had to copy in a hack from an
issue on bootstrap that only half works.

The changes to bootstrap-collapse.js show the hack that I pinched
from https://github.com/twitter/bootstrap/pull/2481 .

The hack gets the outer accordian to expand as needed, but breaks
the expansion animation... I decided this is the better of two evils,
but it is worth re-visiting later, or by someone better at JS than me.

This commit goes most of the way towards #11, but doesn't put the
tabs in MRO.