Change import script to specify import path
Can now specify where a django installation can be found, eg site-packages,
as the argument to the management command, rather than having to make a
django project active in order to populate it.

Still need to work out how to properly reset the environment between loops
of handle_label: if we can do that then we can specify a series of paths and
populate from them all at once, but for now it's just once per run.