Use sqlite database in production
Before now, we've been using a postgres database to store the data in
production. It's not worth the maintenance cost.

This changes the start-up script we use in production to create and
populate an SQLite file before we start the server process.

We probably don't need to take the step of removing the sqlite file, but
I'm not sure if it might already exist from previous builds in prod.