Don't rely on Module.__str__
Before this change, we relied on a string representation of the
Module model, but display logic should not belong on models.

We only needed to get the module name in one place, and use it in two.

We now pass a string to the template instead of a whole model.