Render class list from POPOs, not Models
Before this change, when rendering the `cbv/_klass_list.html` template,
we passed in models and rendered the data from them.

To put a layer between the models and the display logic, this first puts
the required data onto a Plain Old Python Object, and then passes that
into the templates.