Split up requirements into dev and prod files
This breaks up requirements.in into two files.  The dev spec file,
requirements.dev.txt, has a constraint on the compiled prod file so that
we can both avoid duplicating requirements between the files but also
avoid conflicting requirements.

By doing a compile on new files we're going to get brand new versions,
effectively doing an upgrade to everything.  So this commit maintains
the existing requirements.txt purely to constrain the compilation and
avoid this situation.