Update default DB to sqlite
Before this change, we defaulted to running against PostgreSQL,
but that's not what we're running in production.

This changes the default database to be SQLite, to better reflect the
state of production.

Doing so reveals a discrepancy in how the tests run between the two
systems. When we run the snapshot tests against PostgreSQL, the class
methods are ordered in a way that ignored symbols, but when run with
SQLite, the symbols are taken into account.

Given this reflects the current state of production, this is better, so
I've updated the snapshot renders.