Rewrite the README to be clear and succinct
This modernises the project's README to match the current state of the
project (ie dropping tools we should look at), tidies up the prose to be
more succinct, and splits out contribution docs to CONTRIBUTING.md.

The contributor docs have been streamlined to point to the `make`
commands where possible.  I've also added a section from #245 on where
the project is going.  It's a long piece of work so my hope is this sets
expectations without making the maintainers feel bad when folks raise a
PR for defunct or soon-to-be-removed code paths.