Switch to Sphinx's InventoryFile when retrieving URLs
The interface to fetch_inventory changed around the 7-8 major version to
take a Sphinx config object.  However, we're not working with a sphinx
site and just want it to load and parse an inventory file for us.
Luckily, the project ships the InventoryFile class which provides an
interface to do exactly this.