Move method ORM query and objects out of klass detail template
We had a filter that did a complex query at template render time. That
logic has now been moved to the view. Flat objects are introduced to
passed method to the template instead.

The helper function is horrible. We should also improve the method
context data structure, but that would require more work on the
template, so it's out of scope.

Similarly, this makes no effort to add types to the helper function
which was previously a template filter.